http: add evhttp_connection_set_closecb to avoid g_requests hang
Prior to this patch, it was possible that if an RPC was interrupted
before the reply was received, the success callback
evhttp_request_set_on_complete_cb would never be called and
g_requests wouldn't be cleaned up. When attempting to shutdown
bitcoind, it would hang waiting for g_requests.empty().

Co-authored-by: Fabian Jahr <fjahr@protonmail.com>