fuzz: add harness specifically for the compact blocks protocol
This harness does one of five commands in a loop:
- send a cmpctblock message to the test node
- send a blocktxn message to the test node
- send a headers message to the test node
- send a sendcmpct message to the test node
- mine a block

Because valid message structure is guided by the harness, the coverage gets a bit
deeper than the unassisted process_message(s) harnesses.

This test harness is non-deterministic. This is because there is a static TestingSetup
and the chain state is polluted across runs when a valid header is added to the block
index. If a new TestingSetup is made each fuzz iteration, then the executions slow way
down.