test: add MapBlockSourceTest functional test to show mapBlockSource bypass
This introduces a functional test that tests three things related to mapBlockSource:
- Out-of-order valid blocks can be improperly rewarded
- Out-of-order invalid blocks can bypass the BlockChecked-based punishment.
- A stale tip never calls BlockChecked.

The first two sub-tests have to deal with the way that mapBlockSource's emplace(...) and erase(...)
usage cause entries to not be found in the out-of-order block case. The third sub-test asserts that
BlockChecked won't trigger for a stale tip.