log: use std::source_location instead of __FILE__, __LINE__ macros
The std::source_location type stores, among other things, the file
name and line number of a source code location. We get rid of the
macro usage and use std::source_location::current() where applicable.
In a later commit, we use this type as the key type in both an
unordered_map and unordered_set to keep track of rate limiters for
each location.

Co-Authored-By: Niklas Gogge <n.goeggi@gmail.com>