log: introduce LogRateLimiter and SourceLocationCounter classes
LogRateLimiter will be used to keep track of our time-based logging window.
In a later commit, the Logger instance will use a single instance of
LogRateLimiter to determine when the window needs to be reset.

SourceLocationCounter tracks the available bytes left for logging in the
current logging window. Usage will become clear in a later commit. It
does not track the actual source location; it is used as a value in an
unordered_map that is keyed by std::source_location.

Co-Authored-By: Niklas Gogge <n.goeggi@gmail.com>