log: introduce BCLog::LogRateLimiter and BCLog::SourceLocationCounter
LogRateLimiter will be used to keep track of source locations and our
current time-based logging window. It contains an unordered_map and
unordered_set to track source locations. This will be used in a later
commit by BCLog::Logger to manage all of the rate limiting logic.

SourceLocationCounter counts the available bytes left for logging per
source location for the current logging window. It does not track actual
source locations; it is used as a value in m_source_locations.

Co-Authored-By: Niklas Gogge <n.goeggi@gmail.com>